Refactor index.js features into backend-module
